Fix x64 runtime issues across the engine
- xrCore: disable the fixed-size pool allocator on x64 (freelist
  corruption crashed mem_alloc during FS init); add gather_info and
  LogStackTrace for the Debug config; xrDebug.cpp (now the x64 handler)
  gains the g_bug_report_file/g_application_path definitions, the
  std::string fail() overload and logs message arguments.
- xrCore: __asm int 3 -> __debugbreak; ULONG_PTR cast in thread_name.
- xrGame: widen CGraphAbstract::m_edge_count to size_t (CVertex's ctor
  takes size_t*); DWORD_PTR for GetProcessAffinityMask; include the
  stalker animation impl header (forceinline body needed per-TU).
- xrGame/ui: _12b pool element is 3 pointers, not 3 DWORDs (x64 list
  nodes are 24 bytes - every UI window allocation overflowed).
- xrGame: harden GetMaterialByIdx - level geometry references material
  indices past the 60 entries in gamemtl.xr (12312/12316); Win32
  silently read heap garbage, x64 now logs (max 3x) and uses default.
- xrGame: restore dropped Lua aliases (gfind/mod/setn) at script
  startup; log script compile/run errors in Release; revert the
  accidental float-cast of the smart-cover A* limits (float round-trip
  overflowed the u32 sentinels on x64).
- xrEngine: pointer-sized ISpatial_NODE::ptrt; INT_PTR dialog callback;
  SecuROM asm markers become scope-preserving stubs on x64.
- xrCPU_Pipe: portable xrM44_Mul (x87 asm was x86-only); NULL dispatch
  entries for the excluded asm memcopy/memfill units.
- HWCaps: report a single GPU on x64 (2007-era nvapi.lib is x86-only).

+++ engine/xrGame/graph_abstract.h
@@ -45,7 +45,9 @@ public:
 
 private:
 	VERTICES					m_vertices;
-	u32							m_edge_count;
+	// size_t (not u32): CVertex's constructor takes a size_t* to this
+	// counter, which only type-matches on 64-bit if the member is one.
+	size_t						m_edge_count;
 
 public:
 	IC									CGraphAbstract	();
